Changes to match datalackey changes. Build test scripts added. Fixes to reproduce test output on different OSes. Renamed datalackey-fsm to datalackey-state.

+++ test/object/run.sh
@@ -10,7 +10,9 @@ F=
 I=
 for A in $*
 do
-    cp $A a$C
+    if [ -f $A ]; then
+        cp $A a$C
+    fi
     F=$(echo $F a$C $A)
     I=$(echo $I a$C ba$C)
     C=$(expr $C + 1)
